Cleaned up style. Added extra template arugments.
